revert and refactoring and hooks
