+++ package.json
@@ -22,6 +22,7 @@
   "dependencies": {
     "lodash": "^4.17.21",
     "node-sass": "^4.14.1",
+    "prettier": "^2.7.1",
     "react": "^16.12.0",
     "react-dom": "^16.12.0",
     "react-scripts": "^3.2.0",
