+++ src/index.js
@@ -1,8 +1,7 @@
 import React from 'react';
 import { render } from 'react-dom';
 
-// import Counter from './Counter';
-import CounterFunctional from "./CounterFunctional";
+import Counter from './Counter';
 
 import './styles.scss';
 
@@ -10,8 +9,7 @@ const Index = () => {
   return (
     <main className="Application">
       <section className="Counters">
-        {/*<Counter max={-15} step={1} />*/}
-        <CounterFunctional max={10} step={1} />
+        <Counter max={10} step={1} />
       </section>
     </main>
   );
